Deprecate DependencyInjectionNodeObserver, create DependencyInjectionManagerNode and IServicesConfigurator
Deprecates the 'DependencyInjectionNodeObserver' class and introduces the 'DependencyInjectionManagerNode' class and 'IServicesConfigurator' interface. This change splits the role of the 'DependencyInjectionNodeObserver' into two separate components: the 'DependencyInjectionManagerNode' class for managing dependency injection, and the 'IServicesConfigurator' interface for registering services. From now on, dependencies will be registered using all classes implementing the 'IServicesConfigurator' interface with instances in the scene tree at the moment of scene creation.